rename: LocalRoomPhotoService -> RoomPhotoServiceImpl [AIR-128]
S3 용 숙소 사진 서비스만 남기기 위해 rename, 테스트 제거
